Added constants to allow concept of backward move (to replace zeros)
and comments for all the constants to avoid confusion in the future.

ALSO ALLOWED BACKWARD MOVE OF ONE FOOT FOR LONG SHOTS SO THAT WE GO BACK
TO PRE-SENECA RANGE CHECK POST MOVE (i.e. so we can land as close as 7
feet from target even though we need to be at 8 to shoot)

Requires review!